Trading_Chart/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.14.0
pyodbc>=4.0.39
python-dotenv>=1.0.0

# Technical analysis
TA-Lib>=0.4.28
pandas-ta>=0.3.14b0

# GUI
PyQt5>=5.15.9

# Backtesting
backtrader>=1.9.78.123

# Machine learning (for future phases)
scikit-learn>=1.3.0
# tensorflow>=2.13.0  # Uncomment when needed
# torch>=2.0.0  # Uncomment when needed

# Data validation
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0